+ initial update to ts; rename files; create tsconfig, tsconfig.node
