+++ package.json
@@ -23,6 +23,7 @@
     "eslint-plugin-react": "^7.34.2",
     "eslint-plugin-react-hooks": "^4.6.2",
     "eslint-plugin-react-refresh": "^0.4.7",
+    "typescript": "^5.5.3",
     "vite": "^5.3.1"
   }
 }
