+++ src/components/TripCard.tsx
