+++ src/components/TripModal.tsx
