+++ src/pages/TripDetails.tsx
